Fix Travis CI Build Errors
Errors are from recent updates to WordPress Coding Standards. Majority of the errors are related to missing translation comments for placeholders such as '%s'.
