+++ functions.php
@@ -20,11 +20,15 @@ require_once( 'library/cleanup.php' );
 /** Required for Foundation to work properly */
 require_once( 'library/foundation.php' );
 
+/** Format comments */
+require_once( 'library/class-foundationpress-comments.php' );
+
 /** Register all navigation menus */
 require_once( 'library/navigation.php' );
 
 /** Add menu walkers for top-bar and off-canvas */
-require_once( 'library/menu-walkers.php' );
+require_once( 'library/class-foundationpress-top-bar-walker.php' );
+require_once( 'library/class-foundationpress-mobile-walker.php' );
 
 /** Create widget areas in sidebar and footer */
 require_once( 'library/widget-areas.php' );
@@ -48,4 +52,4 @@ require_once( 'library/sticky-posts.php' );
 require_once( 'library/responsive-images.php' );
 
 /** If your site requires protocol relative url's for theme assets, uncomment the line below */
-// require_once( 'library/protocol-relative-theme-assets.php' );
+// require_once( 'library/class-foundationpress-protocol-relative-theme-assets.php' );
