+++ library/responsive-images.php
@@ -41,13 +41,21 @@ function foundationpress_adjust_image_sizes_attr( $sizes, $size ) {
 
 	// Full width page template
 	if ( is_page_template( 'page-templates/page-full-width.php' ) ) {
-		1200 < $width && $sizes = '(max-width: 1199px) 98vw, 1200px';
-		1200 > $width && $sizes = '(max-width: 1199px) 98vw, ' . $width . 'px';
+		if ( 1200 < $width ) {
+			$sizes = '(max-width: 1199px) 98vw, 1200px';
+		}
+		if ( 1200 > $width ) {
+			$sizes = '(max-width: 1199px) 98vw, ' . $width . 'px';
+		}
 
 	// Default 3/4 column post/page layout
 	} else {
-		770 < $width && $sizes = '(max-width: 639px) 98vw, (max-width: 1199px) 64vw, 770px';
-		770 > $width && $sizes = '(max-width: 639px) 98vw, (max-width: 1199px) 64vw, ' . $width . 'px';
+		if ( 770 < $width ) {
+			$sizes = '(max-width: 639px) 98vw, (max-width: 1199px) 64vw, 770px';
+		}
+		if ( 770 > $width ) {
+			$sizes = '(max-width: 639px) 98vw, (max-width: 1199px) 64vw, ' . $width . 'px';
+		}
 	}
 
 	return $sizes;
