+++ page.php
@@ -28,8 +28,15 @@
            <?php edit_post_link( __( 'Edit', 'foundationpress' ), '<span class="edit-link">', '</span>' ); ?>
        </div>
        <footer>
-           <?php wp_link_pages( array('before' => '<nav id="page-nav"><p>' . __( 'Pages:', 'foundationpress' ), 'after' => '</p></nav>' ) ); ?>
-           <p><?php the_tags(); ?></p>
+          <?php
+            wp_link_pages(
+              array(
+                'before' => '<nav id="page-nav"><p>' . __( 'Pages:', 'foundationpress' ),
+                'after'  => '</p></nav>',
+              )
+            );
+          ?>
+          <p><?php the_tags(); ?></p>
        </footer>
        <?php do_action( 'foundationpress_page_before_comments' ); ?>
        <?php comments_template(); ?>
