+++ single.php
@@ -25,7 +25,14 @@ get_header(); ?>
 			<?php edit_post_link( __( 'Edit', 'foundationpress' ), '<span class="edit-link">', '</span>' ); ?>
 		</div>
 		<footer>
-			<?php wp_link_pages( array('before' => '<nav id="page-nav"><p>' . __( 'Pages:', 'foundationpress' ), 'after' => '</p></nav>' ) ); ?>
+			<?php
+				wp_link_pages(
+					array(
+						'before' => '<nav id="page-nav"><p>' . __( 'Pages:', 'foundationpress' ),
+						'after'  => '</p></nav>',
+					)
+				);
+			?>
 			<p><?php the_tags(); ?></p>
 		</footer>
 		<?php the_post_navigation(); ?>
