+++ template-parts/content-none.php
@@ -17,7 +17,15 @@
 <div class="page-content">
 	<?php if ( is_home() && current_user_can( 'publish_posts' ) ) : ?>
 
-	<p><?php printf( __( 'Ready to publish your first post? <a href="%1$s">Get started here</a>.', 'foundationpress' ), admin_url( 'post-new.php' ) ); ?></p>
+	<p>
+		<?php
+			/* translators: %1$s: new post url */
+			printf( __(
+				'Ready to publish your first post? <a href="%1$s">Get started here</a>.', 'foundationpress' ),
+				admin_url( 'post-new.php' )
+			);
+		?>
+	</p>
 
 	<?php elseif ( is_search() ) : ?>
